aetherbound/requirements.txt
PyOpenGL
glfw
imgui[glfw]
PyGLM
numpy
pygame
